requirements.txt
ftfy==6.1.1
gdown==4.7.1
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
protobuf==3.9.2
PyYAML==6.0
scipy==1.10.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1+cu116
torch-tb-profiler==0.4.0
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.65.0
cupy-cuda112